+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -20,8 +20,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
-constexpr std::uint16_t translation_table_mapping_table_size = 415;
-constexpr std::size_t translation_table_string_table_size = 75181;
+constexpr std::uint16_t translation_table_mapping_table_size = 418;
+constexpr std::size_t translation_table_string_table_size = 75300;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -59,6 +59,8 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "'declare class' methods cannot be marked 'async'"sv,
           "'declare class' methods cannot be marked as a generator"sv,
           "'declare class' methods cannot contain a body"sv,
+          "'declare {1}' cannot have initializer"sv,
+          "'declare {1}' started here"sv,
           "'do-while' loop"sv,
           "'else' has no corresponding 'if'"sv,
           "'extends' must be before 'implements'"sv,
@@ -97,6 +99,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "TypeScript 'as' type assertions are not allowed in JavaScript"sv,
           "TypeScript 'declare abstract class' is not allowed in JavaScript"sv,
           "TypeScript 'declare class' is not allowed in JavaScript"sv,
+          "TypeScript 'declare {1}' is not allowed in JavaScript"sv,
           "TypeScript 'implements' is not allowed in JavaScript"sv,
           "TypeScript <Type> type assertions are not allowed in JSX mode"sv,
           "TypeScript assignment-asserted fields are not supported in JavaScript"sv,